Renaming class, separating enum, importing enum, enhanced switch přejmenování classy, oddělení enum, import enum, vylepšený swtich
