+++ src/main/com/example006/radek/Operation.java
@@ -1,0 +1,8 @@
+package main.com.example006.radek;
+
+public enum Operation {
+        PLUS,
+        MINUS,
+        MULTI,
+        DIVIDE
+}
